+++ tests/unit/pywbemcli/test_help_cmd.py
@@ -15,7 +15,7 @@
 # limitations under the License.
 """
 Test the help command that displays help text on specific pywbemcli
-subjects.
+subjects. This command is not part of any command group.
 
 """
 
@@ -25,10 +25,13 @@ from .cli_test_extensions import CLITestsBase
 
 DEFAULT_HELP_LINES = """Help subjects
 subject name    subject description
---------------  --------------------------------------------
+--------------  ---------------------------------------------
+activate        Activating shell tab completion
 instancename    InstanceName parameter in instance cmd group
 repl            Using the repl command
+tab-completion  Where tab completion is provided by pywbemcli
 """
+
 REPL_HELP_LINES = [
     'repl - Using the repl command',
     'In the interactive mode pywbem returns control to a terminal. General'
@@ -39,6 +42,15 @@ INSTANCENAME_HELP_LINES = [
     "An instance path is specified using the INSTANCENAME argument and "
 ]
 
+TABCOMPLETION_HELP_LINES = [
+    "Tab completion is always available in the interactive mode (see help repl)"
+]
+
+ACTIVATE_HELP_LINES = [
+    "Pywbemcli includes tab-completion capability for all commands for certain"
+]
+
+
 OK = True     # mark tests OK when they execute correctly
 RUN = True    # Mark OK = False and current test case being created RUN
 FAIL = False  # Any test currently FAILING or not tested yet
@@ -46,7 +58,7 @@ SKIP = False
 
 
 TEST_CASES = [
-    # List of testcases.
+    # List of testcases for test help <subject>.
     # Each testcase is a list with the following items:
     # * desc: Description of testcase.
     # * inputs: String, or tuple/list of strings, or dict of 'env', 'args',
@@ -63,46 +75,58 @@ TEST_CASES = [
     #     is skipped.
 
     ['Verify help command default response list of subjects',
-     {'subject': []},
+     {'args': ['help'], },
      {'stdout': DEFAULT_HELP_LINES,
       'test': 'innows'},
-     None, OK],
+     None, RUN],
 
     ['Verify help command repl',
-     {'subject': ['repl']},
+     {'args': ['help', 'repl']},
      {'stdout': REPL_HELP_LINES,
       'test': 'innows'},
      None, OK],
 
     ['Verify help command instancename',
-     {'subject': ['instancename']},
+     {'args': ['help', 'instancename']},
      {'stdout': INSTANCENAME_HELP_LINES,
       'test': 'innows'},
      None, OK],
+
+    ['Verify help command tab-completion',
+     {'args': ['help', 'tab-completion']},
+     {'stdout': TABCOMPLETION_HELP_LINES,
+      'test': 'innows'},
+     None, OK],
+
+    ['Verify help command tab-completion',
+     {'args': ['help', 'blah']},
+     {'stderr': ["'blah' is not a valid help subject."],
+      'rc': 1,
+      'test': 'innows'},
+     None, OK],
+
 ]
 
 
-class TestHelpCmd(CLITestsBase):  # pylint: disable=too-few-pubic-methods
+class TestCmdHelp(CLITestsBase):  # pylint: disable=too-few-public-methods
     """
     Test the general options including statistics,  --server,
     --timeout, --use-pull, --pull-max-cnt, --output-format
     """
+
+    command_group = None
+
     @pytest.mark.parametrize(
         "desc, inputs, exp_response, mock, condition", TEST_CASES)
-    def test_execute_pywbemcli(self, desc, inputs, exp_response, mock,
-                               condition):
-        # pylint: disable=unused-argument
+    def test_help_cmd(self, desc, inputs, exp_response, mock, condition):
         """
-        Execute pybemcli with the defined input and test output.
+        Common test method for those commands and options in the
+        class command that can be tested.  This includes:
 
+          * Subcommands like help that do not require access to a server
+
+          * Subcommands that can be tested with a single execution of a
+            pywbemcli command.
         """
-
-        command_group = 'help'
-
-        if inputs['subject']:
-            inputs = inputs['subject']
-        else:
-            inputs = []
-
-        self.command_test(desc, command_group, inputs, exp_response,
-                          None, condition)
+        self.command_test(desc, self.command_group, inputs, exp_response,
+                          mock, condition, verbose=True)